[Enterprise Search] Set up cypress-axe tests (#108465)
* Set up cypress-axe

@see https://github.com/component-driven/cypress-axe

* DRY out Kibana axe rules into constants that Cypress can use

* Create shared & configured checkA11y command

+ fix string union type error
+ remove unnecessary tsconfig exclude

* Add Overview plugin a11y tests

* Add AS & WS placeholder a11y checks

- Mostly just re-exporting the shared command and checking for failures, I only ran this after the shared axe config settings and found no failures

* Configure our axe settings further to catch best practices

- notably heading level issues (thanks Byron for catching this!)

- however I now also need to set an ignore on a duplicate landmark violation caused by the global header (not sure why it's showing up - shouldn't it be out of context? bah)

- remove option to pass args into checkA11y - I figure it's not super likely we'll need to override axe settings per-page (vs not running it), but we can pass it custom configs or args later if needed
