+++ x-pack/plugins/enterprise_search/public/applications/shared/cypress/tsconfig.json
@@ -1,8 +1,9 @@
 {
   "extends": "../../../../../../../tsconfig.base.json",
+  "references": [{ "path": "../../../../../../../test/tsconfig.json" }],
   "include": ["./**/*"],
   "compilerOptions": {
     "outDir": "../../../../target/cypress/types/shared",
-    "types": ["cypress", "node"]
+    "types": ["cypress", "cypress-axe", "node"]
   }
 }
